creating dedicated activity for image selection
