+++ features/home/src/main/AndroidManifest.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android" package="app.dapk.st.home">
 
-  <application>
-    <activity android:name="app.dapk.st.home.MainActivity"/>
-  </application>
+    <application>
+        <activity android:name="app.dapk.st.home.MainActivity"/>
+        <activity android:name="app.dapk.st.home.gallery.ImageGalleryActivity"/>
+    </application>
 
 </manifest>
