+++ features/home/src/main/kotlin/app/dapk/st/home/gallery/ImageGalleryScreen.kt
@@ -1,0 +1,160 @@
+package app.dapk.st.home
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import app.dapk.st.core.Lce
+import app.dapk.st.core.LifecycleEffect
+import app.dapk.st.core.components.CenteredLoading
+import app.dapk.st.design.components.GenericError
+import app.dapk.st.design.components.Spider
+import app.dapk.st.design.components.SpiderPage
+import app.dapk.st.home.gallery.Folder
+import app.dapk.st.home.gallery.ImageGalleryViewModel
+import app.dapk.st.home.gallery.Media
+import coil.compose.rememberAsyncImagePainter
+import coil.request.ImageRequest
+
+@Composable
+fun ImageGalleryScreen(viewModel: ImageGalleryViewModel, onTopLevelBack: () -> Unit, onImageSelected: (Media) -> Unit) {
+    LifecycleEffect(onStart = {
+        viewModel.start()
+    })
+
+    val onNavigate: (SpiderPage<out ImageGalleryPage>?) -> Unit = {
+        when (it) {
+            null -> onTopLevelBack()
+            else -> viewModel.goTo(it)
+        }
+    }
+
+    Spider(currentPage = viewModel.state.page, onNavigate = onNavigate) {
+        item(ImageGalleryPage.Routes.folders) {
+            ImageGalleryFolders(it) { folder ->
+                viewModel.selectFolder(folder)
+            }
+        }
+        item(ImageGalleryPage.Routes.files) {
+            ImageGalleryMedia(it, onImageSelected)
+        }
+    }
+
+}
+
+
+@Composable
+fun ImageGalleryFolders(state: ImageGalleryPage.Folders, onClick: (Folder) -> Unit) {
+    val screenWidth = LocalConfiguration.current.screenWidthDp
+
+    val gradient = Brush.verticalGradient(
+        colors = listOf(Color.Transparent, Color.Black.copy(alpha = 0.5f)),
+    )
+
+    when (val content = state.content) {
+        is Lce.Loading -> {
+            CenteredLoading()
+        }
+
+        is Lce.Content -> {
+            Column {
+                val columns = when {
+                    screenWidth > 600 -> 4
+                    else -> 2
+                }
+                LazyVerticalGrid(
+                    columns = GridCells.Fixed(columns),
+                    modifier = Modifier.fillMaxSize(),
+                ) {
+                    items(content.value, key = { it.bucketId }) {
+                        Box(modifier = Modifier.fillMaxWidth().padding(2.dp).aspectRatio(1f)
+                            .clickable { onClick(it) }) {
+                            Image(
+                                painter = rememberAsyncImagePainter(
+                                    model = ImageRequest.Builder(LocalContext.current)
+                                        .data(it.thumbnail.toString())
+                                        .build(),
+                                ),
+                                contentDescription = "123",
+                                modifier = Modifier.fillMaxSize(),
+                                contentScale = ContentScale.Crop
+                            )
+
+                            Box(modifier = Modifier.fillMaxWidth().fillMaxHeight(0.6f).background(gradient).align(Alignment.BottomStart))
+                            Row(
+                                modifier = Modifier.fillMaxWidth().align(Alignment.BottomStart).padding(4.dp),
+                                horizontalArrangement = Arrangement.SpaceBetween,
+                                verticalAlignment = Alignment.CenterVertically
+                            ) {
+                                Text(it.title, fontSize = 13.sp, color = Color.White)
+                                Text(it.itemCount.toString(), fontSize = 11.sp, color = Color.White)
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+        is Lce.Error -> GenericError { }
+    }
+}
+
+@Composable
+fun ImageGalleryMedia(state: ImageGalleryPage.Files, onFileSelected: (Media) -> Unit) {
+    val screenWidth = LocalConfiguration.current.screenWidthDp
+
+    Column {
+        val columns = when {
+            screenWidth > 600 -> 4
+            else -> 2
+        }
+
+        when (val content = state.content) {
+            is Lce.Loading -> {
+                CenteredLoading()
+            }
+
+            is Lce.Content -> {
+                LazyVerticalGrid(
+                    columns = GridCells.Fixed(columns),
+                    modifier = Modifier.fillMaxSize(),
+                ) {
+                    val modifier = Modifier.fillMaxWidth().padding(2.dp).aspectRatio(1f)
+                    items(content.value, key = { it.id }) {
+                        Box(modifier = modifier.clickable { onFileSelected(it) }) {
+                            Image(
+                                painter = rememberAsyncImagePainter(
+                                    model = ImageRequest.Builder(LocalContext.current)
+                                        .data(it.uri.toString())
+                                        .crossfade(true)
+                                        .build(),
+                                ),
+                                contentDescription = "123",
+                                modifier = Modifier.fillMaxWidth().fillMaxHeight(),
+                                contentScale = ContentScale.Crop
+                            )
+                        }
+                    }
+                }
+            }
+
+            is Lce.Error -> GenericError { }
+        }
+
+    }
+
+}
