+++ features/navigator/src/main/kotlin/app/dapk/st/navigator/Navigator.kt
@@ -1,6 +1,7 @@
 package app.dapk.st.navigator
 
 import android.app.Activity
+import android.app.Instrumentation.ActivityResult
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
